+++ source-files/web-components/help-panel.js
@@ -89,7 +89,7 @@ window.customElements.define('help-panel', class extends HTMLElement {
           }
         }
       } else {
-        this.hidePanel();
+        this.collapsePanel();
       }
     });
 
@@ -111,6 +111,19 @@ window.customElements.define('help-panel', class extends HTMLElement {
         licenseDivEl.classList.add('global-text-theme-dark');
         infoDivEl.classList.remove('global-text-theme-light');
         infoDivEl.classList.add('global-text-theme-dark');
+      }
+    });
+
+    /**
+     * Global event listener on document object to detect state change of remote IRC server
+     * Detect addition of new IRC channels and create channel panel.
+     * Data source: ircState object
+     * @listens document:irc-state-changed
+     */
+    document.addEventListener('irc-state-changed', () => {
+      if (window.globals.ircState.ircConnected !== this.ircConnectedLast) {
+        this.ircConnectedLast = window.globals.ircState.ircConnected;
+        if (!window.globals.ircState.ircConnected) this.hidePanel();
       }
     });
 
@@ -155,7 +168,7 @@ window.customElements.define('help-panel', class extends HTMLElement {
           }
         }
       } else {
-        if (event.detail.debug) this.showPanel();
+        if ((event.detail) && (event.detail.debug)) this.showPanel();
       }
     });
   } // connectedCallback()