+++ source-files/web-components/logout-panel.js
@@ -110,7 +110,7 @@ window.customElements.define('logout-panel', class extends HTMLElement {
           }
         }
       } else {
-        this.hidePanel();
+        this.collapsePanel();
       }
     });
 
@@ -155,7 +155,7 @@ window.customElements.define('logout-panel', class extends HTMLElement {
           }
         }
       } else {
-        if (event.detail.debug) this.showPanel();
+        if ((event.detail) && (event.detail.debug)) this.showPanel();
       }
     });
   } // connectedCallback()