+++ source-files/web-components/server-form.js
@@ -87,7 +87,7 @@ window.customElements.define('server-form', class extends HTMLElement {
           }
         }
       } else {
-        this.hidePanel();
+        this.collapsePanel();
       }
     });
 
@@ -143,7 +143,7 @@ window.customElements.define('server-form', class extends HTMLElement {
           }
         }
       } else {
-        if (event.detail.debug) this.showPanel();
+        if ((event.detail) && (event.detail.debug)) this.showPanel();
       }
     });
   } // connectedCallback()