+++ source-files/web-components/show-ircstate.js
@@ -105,6 +105,19 @@ window.customElements.define('show-ircstate', class extends HTMLElement {
           }
         }
       } else {
+        this.collapsePanel();
+      }
+    });
+
+    /**
+     * Global event listener on document object to detect state change of remote IRC server
+     * Detect addition of new IRC channels and create channel panel.
+     * Data source: ircState object
+     * @listens document:irc-state-changed
+     */
+    document.addEventListener('irc-state-changed', () => {
+      if (window.globals.ircState.ircConnected !== this.ircConnectedLast) {
+        this.ircConnectedLast = window.globals.ircState.ircConnected;
         this.hidePanel();
       }
     });
@@ -170,7 +183,7 @@ window.customElements.define('show-ircstate', class extends HTMLElement {
           }
         }
       } else {
-        if (event.detail.debug) this.showPanel();
+        if ((event.detail) && (event.detail.debug)) this.showPanel();
       }
     });
   } // connectedCallback()