+++ source-files/web-components/show-webstate.js
@@ -105,7 +105,7 @@ window.customElements.define('show-webstate', class extends HTMLElement {
           }
         }
       } else {
-        this.hidePanel();
+        this.collapsePanel();
       }
     });
 
@@ -129,6 +129,20 @@ window.customElements.define('show-webstate', class extends HTMLElement {
         jsonDisplayDivEl.classList.add('global-text-theme-dark');
       }
     }); // color-theme-changed
+
+    /**
+     * Global event listener on document object to detect state change of remote IRC server
+     * Detect addition of new IRC channels and create channel panel.
+     * Data source: ircState object
+     * @listens document:irc-state-changed
+     */
+    document.addEventListener('irc-state-changed', () => {
+      if (window.globals.ircState.ircConnected !== this.ircConnectedLast) {
+        this.ircConnectedLast = window.globals.ircState.ircConnected;
+        if (!window.globals.ircState.ircConnected) this.hidePanel();
+      }
+    });
+
     /**
      * Hide panel (not visible) unless listed as exception.
      * @listens document:hide-all-panels
@@ -170,7 +184,7 @@ window.customElements.define('show-webstate', class extends HTMLElement {
           }
         }
       } else {
-        if (event.detail.debug) this.showPanel();
+        if ((event.detail) && (event.detail.debug)) this.showPanel();
       }
     });
   } // connectedCallback()